HP bar attached to damage + Leaper damage works!

+++ Production Project/Assets/_OurAssets/_Scripts/EnemyScripts/EnemyTest/Unit.cs
@@ -6,9 +6,9 @@ using UnityEngine.UI;
 public class Unit : MonoBehaviour
 {
 
-    [SerializeField] internal int _unitHP;
-    [SerializeField] internal int _unitMaxHP;
-    [SerializeField] internal int _unitRange;
+    [SerializeField] internal float _unitHP;
+    [SerializeField] internal float _unitMaxHP;
+    [SerializeField] internal float _unitRange;
 
     [SerializeField] Slider _myHealthSlider;
